test: database mocking configured bug: modifed some queries to be findOne instead of findAll

+++ server.js
@@ -13,9 +13,11 @@ const db = require('./src/config/db.config.js');
 global.__basedir = __dirname;
   
 // Resets and clears the database when the application is started
-db.sequelize.sync({force: true}).then(() => {
-  console.log('\n', '\n', '#### Resetting database and clearing all data. Ready to accept requests! ####', '\n', '\n');
-}); 
+if (db.sequelize.options.dialect !== 'mock') {
+  db.sequelize.sync({force: true}).then(() => {
+    console.log('\n', '\n', '#### Resetting database and clearing all data. Ready to accept requests! ####', '\n', '\n');
+  }); 
+}
  
 // Router connections
 let imageRouter = require('./src/routers/image.router.js');
@@ -31,3 +33,5 @@ const server = app.listen(8080, function () {
  
   console.log('\n', '\n', 'App listening at http://' + host + ':' + port, '\n', '\n'); 
 });
+
+module.exports = server;
